requirements.txt
flask==3.0.0
langchain==0.0.350
openai==0.28.1
python-dotenv==1.0.0
SQLAlchemy==2.0.27
aiohttp==3.9.3
async-timeout==4.0.3
requests==2.31.0
typing-extensions==4.9.0
pydantic==1.10.13
numpy==1.26.3
tenacity==8.2.3 